add error boundary + move LoadingPage to misc
