+++ package.json
@@ -20,6 +20,7 @@
         "react": "^18.2.0",
         "react-datepicker": "^4.10.0",
         "react-dom": "^18.2.0",
+        "react-error-boundary": "^4.0.9",
         "react-hook-form": "^7.43.9",
         "react-modal": "^3.16.1",
         "react-router-dom": "^6.6.1",
