+++ src/features/misc/ErrorFallback/ErrorFallback.module.scss
@@ -1,0 +1,24 @@
+@import "../../../assets/variables";
+
+.page {
+    width: 100vw;
+    height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: $bg-dark;
+    color: $text-dark;
+    font-size: 1.5rem;
+    .card {
+        display: flex;
+        flex-direction: column;
+        gap: 1rem;
+        align-items: center;
+        h1 {
+            margin-bottom: 1rem;
+        }
+        span {
+            display: block;
+        }
+    }
+}
