+++ src/features/misc/ErrorFallback/index.ts
@@ -1,0 +1,1 @@
+export { default as ErrorFallback } from "./ErrorFallback";
