+++ src/features/misc/LoadingPage/LoadingPage.module.scss
@@ -1,4 +1,4 @@
-@import "../../assets/variables";
+@import "../../../assets/variables";
 
 .page {
     width: 100vw;
