+++ src/providers/MainProviders.tsx
@@ -1,10 +1,12 @@
 import { QueryClientProvider } from "@tanstack/react-query";
+import { ErrorBoundary } from "react-error-boundary";
 import { Suspense } from "react";
 import { BrowserRouter } from "react-router-dom";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import queryClient from "@/lib/react-query";
 import { ThemeProvider } from "@/context/ThemeContext";
-import { LoadingPage } from "@/features/LoadingPage";
+import { LoadingPage } from "@/features/misc/LoadingPage";
+import { ErrorFallback } from "@/features/misc/ErrorFallback";
 
 type MainProvidersProps = {
     children: React.ReactNode;
@@ -13,12 +15,14 @@ type MainProvidersProps = {
 function MainProviders({ children }: MainProvidersProps) {
     return (
         <ThemeProvider>
-            <Suspense fallback={<LoadingPage />}>
-                <QueryClientProvider client={queryClient}>
-                    {import.meta.env.DEV && <ReactQueryDevtools />}
-                    <BrowserRouter>{children}</BrowserRouter>
-                </QueryClientProvider>
-            </Suspense>
+            <ErrorBoundary fallbackRender={ErrorFallback}>
+                <Suspense fallback={<LoadingPage />}>
+                    <QueryClientProvider client={queryClient}>
+                        {import.meta.env.DEV && <ReactQueryDevtools />}
+                        <BrowserRouter>{children}</BrowserRouter>
+                    </QueryClientProvider>
+                </Suspense>
+            </ErrorBoundary>
         </ThemeProvider>
     );
 }
